requirements.txt
Django==1.6.1
Jinja2==2.7.2
Markdown==2.3.1
MarkupSafe==0.18
MySQL-python==1.2.5
South==0.8.4
amqp==1.4.1
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.13
bitcoin-python==0.3
celery==3.1.7
django-bootstrap-toolkit==2.15.0
django-bootstrap3==3.0.0
django-filter==0.7
django-jinja==0.23.1
django-widget-tweaks==1.3
djangorestframework==2.3.12
gunicorn==18.0
kombu==3.0.9
meld3==0.6.10
python-jsonrpc==0.3.4
pytz==2013.9
wsgiref==0.1.2
